Intento de arreglo, codigo de verificacion
