+++ backend/controllers/userController.js
@@ -1,8 +1,9 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
 import dotenv from "dotenv";
-import User, { validarEmail, validarDNI } from "../models/User.js";
-import nodemailer from "nodemailer";
+import User from "../models/User.js";
+import { generateVerificationCode } from "../utils/verification.js";
+import { sendVerificationEmail } from "../utils/emailService.js";
 
 dotenv.config();
 
@@ -13,10 +14,10 @@ function generateVerificationCode() {
 
 //Configuración de nodemailer para enviar correos
 const transporter = nodemailer.createTransport({
-  service: "gmail", 
+  service: "gmail",
   auth: {
-    user: process.env.EMAIL_USER, 
-    pass: process.env.EMAIL_PASS, 
+    user: process.env.EMAIL_USER,
+    pass: process.env.EMAIL_PASS,
   },
 });
 
@@ -30,21 +31,7 @@ export const registerUser = async (req, res) => {
     const { name, surname, email, password, dni, grade } = req.body;
 
     if (!name || !surname || !email || !password || !dni || !grade) {
-      return res
-        .status(400)
-        .json({ mensaje: "Todos los campos son obligatorios" });
-    }
-
-    if (!validarEmail(email)) {
-      return res
-        .status(400)
-        .json({ mensaje: "Solo se permiten correos de U-TAD." });
-    }
-
-    if (!validarDNI(dni)) {
-      return res
-        .status(400)
-        .json({ mensaje: "DNI inválido. Debe seguir el formato correcto." });
+      return res.status(400).json({ mensaje: "Todos los campos son obligatorios" });
     }
 
     const usuarioExistente = await User.findOne({ email });
@@ -57,7 +44,6 @@ export const registerUser = async (req, res) => {
       return res.status(400).json({ mensaje: "El DNI ya está registrado." });
     }
 
-    const verificationCode = generateVerificationCode(); //Generar código de verificación
     const salt = await bcrypt.genSalt(10);
     const passwordHasheada = await bcrypt.hash(password, salt);
 
@@ -68,61 +54,105 @@ export const registerUser = async (req, res) => {
       password: passwordHasheada,
       dni,
       grade,
-      rol: "user", // Forzamos el rol para evitar registros no autorizados
-      isVerified: false, //Usuario no verificado aún
-      verificationCode,
-      verificationAttempts: 3,
+      rol: "user",
+      isVerified: true, // 🔥 Se deja como verificado temporalmente para permitir login
+      // verificationCode: generateVerificationCode(),  // ❌ Comentar la generación del código
+      // verificationAttempts: 3,
+      // verificationCodeExpires: new Date(Date.now() + 10 * 60 * 1000), // Expiración en 10 min
     });
 
     await nuevoUsuario.save();
 
-    //Enviar correo con el código de verificación
-    const mailOptions = {
-      from: `"Bildy" <${process.env.EMAIL_USER}>`,
-      to: email,
-      subject: "Código de Verificación",
-      text: `Tu código de verificación es: ${verificationCode}`,
-    };
+    // ❌ Comentar el envío de correo hasta la siguiente iteración
+    // await sendVerificationEmail(email, verificationCode);
 
-    await transporter.sendMail(mailOptions);
-
-    res.status(201).json({ mensaje: "Usuario registrado exitosamente." });
+    res.status(201).json({ mensaje: "Usuario registrado con éxito." });
   } catch (error) {
     console.error("❌ Error en el servidor:", error);
     res.status(500).json({ mensaje: "Error en el servidor." });
   }
 };
 
-//Verificar código de verificación
-export const verifyCode = async (req, res) => {
-  try {
-    const { email, code } = req.body;
-    const user = await User.findOne({ email });
+// export const verifyCode = async (req, res) => {
+//   try {
+//     const { email, code } = req.body;
+//     const user = await User.findOne({ email });
 
-    if (!user) {
-      return res.status(404).json({ mensaje: "Usuario no encontrado." });
-    }
+//     if (!user) {
+//       return res.status(404).json({ mensaje: "Usuario no encontrado." });
+//     }
 
-    if (user.verificationAttempts <= 0) {
-      return res.status(403).json({ mensaje: "Has agotado tus intentos de verificación." });
-    }
+//     if (user.isVerified) {
+//       return res
+//         .status(400)
+//         .json({ mensaje: "Este usuario ya está verificado." });
+//     }
 
-    if (user.verificationCode === code) {
-      user.isVerified = true;
-      user.verificationCode = null;
-      await user.save();
-      return res.json({ mensaje: "Código correcto, usuario verificado." });
-    } else {
-      user.verificationAttempts -= 1;
-      await user.save();
-      return res.status(400).json({ mensaje: `Código incorrecto. Intentos restantes: ${user.verificationAttempts}` });
-    }
-  } catch (error) {
-    console.error("❌ Error en verifyCode:", error);
-    res.status(500).json({ mensaje: "Error en el servidor." });
-  }
-};
+//     if (new Date() > user.verificationCodeExpires) {
+//       user.verificationCode = generateVerificationCode();
+//       user.verificationAttempts = 3;
+//       user.verificationCodeExpires = new Date();
+//       user.verificationCodeExpires.setMinutes(
+//         user.verificationCodeExpires.getMinutes() + 10
+//       );
+//       await user.save();
+//       await sendVerificationEmail(email, user.verificationCode);
 
+//       return res
+//         .status(400)
+//         .json({ mensaje: "El código expiró. Se ha enviado uno nuevo." });
+//     }
+
+//     if (user.verificationCode === code) {
+//       user.isVerified = true;
+//       user.verificationCode = null;
+//       await user.save();
+//       return res.json({ mensaje: "Código correcto, usuario verificado." });
+//     } else {
+//       user.verificationAttempts -= 1;
+//       await user.save();
+//       return res.status(400).json({
+//         mensaje: `Código incorrecto. Intentos restantes: ${user.verificationAttempts}`,
+//       });
+//     }
+//   } catch (error) {
+//     console.error("❌ Error en verifyCode:", error);
+//     res.status(500).json({ mensaje: "Error en el servidor." });
+//   }
+// };
+
+// // Reenviar código de verificación
+// export const resendVerificationCode = async (req, res) => {
+//   try {
+//     const { email } = req.body;
+//     const user = await User.findOne({ email });
+
+//     if (!user) {
+//       return res.status(404).json({ mensaje: "Usuario no encontrado." });
+//     }
+
+//     if (user.isVerified) {
+//       return res
+//         .status(400)
+//         .json({ mensaje: "Este usuario ya está verificado." });
+//     }
+
+//     user.verificationCode = generateVerificationCode();
+//     user.verificationCodeExpires = new Date();
+//     user.verificationCodeExpires.setMinutes(
+//       user.verificationCodeExpires.getMinutes() + 10
+//     );
+//     user.verificationAttempts = 3;
+
+//     await user.save();
+//     await sendVerificationEmail(email, user.verificationCode);
+
+//     res.json({ mensaje: "Se ha enviado un nuevo código de verificación." });
+//   } catch (error) {
+//     console.error("❌ Error en resendVerificationCode:", error);
+//     res.status(500).json({ mensaje: "Error en el servidor." });
+//   }
+// };
 
 /**
  * @desc Iniciar sesión y obtener un token JWT
@@ -218,7 +248,7 @@ export const getUserProfile = async (req, res) => {
       id: usuario._id,
       name: usuario.name,
       email: usuario.email,
-      rol: usuario.rol, 
+      rol: usuario.rol,
     });
   } catch (error) {
     console.error("❌ Error en el servidor:", error);
