+++ backend/models/User.js
@@ -1,26 +1,5 @@
 import mongoose from "mongoose";
-
-// Función para calcular la letra correcta del DNI
-export const calcularLetraDNI = (dni) => {
-  const letras = "TRWAGMYFPDXBNJZSQVHLCKE";
-  return letras[dni % 23];
-};
-
-// Función de validación del DNI
-export const validarDNI = (dni) => {
-  const regex = /^\d{8}[A-Z]$/;
-  if (!regex.test(dni)) return false;
-
-  const numeros = parseInt(dni.slice(0, 8), 10);
-  const letra = dni.charAt(8);
-
-  return calcularLetraDNI(numeros) === letra;
-};
-
-// Función de validación de email
-export const validarEmail = (email) => {
-  return /@u-tad\.com$|@live\.u-tad\.com$/.test(email);
-};
+import { validarEmail, validarDNI } from "../utils/validators.js"; // Importamos validaciones externas
 
 const userSchema = new mongoose.Schema({
   name: { type: String, required: true },
@@ -57,11 +36,11 @@ const userSchema = new mongoose.Schema({
   },
   projects: [{ type: mongoose.Schema.Types.ObjectId, ref: "Project" }], //relacion inversa
 
-  //Campos para la verificación del código de 6 dígitos
+  // Campos para la verificación del código de 6 dígitos
   isVerified: { type: Boolean, default: false }, // Indica si el usuario ya verificó su cuenta
   verificationCode: { type: String, default: null }, // Código de verificación temporal
-  verificationAttempts: { type: Number, default: 3 }, // Número de intentos para ingresar el código
- 
+  verificationAttempts: { type: Number, default: 3 }, // Número de intentos
+  verificationCodeExpires: { type: Date }, // Expiración del código
 });
 
 const User = mongoose.model("User", userSchema);
